Reorganize dialog views into a dialogs package
