+++ src/vorta/borg/extract.py
@@ -3,7 +3,7 @@ import tempfile
 from PyQt6.QtCore import QModelIndex, Qt
 
 from vorta.utils import borg_compat
-from vorta.views.extract_dialog import ExtractTree, FileData
+from vorta.views.dialogs.extract_dialog import ExtractTree, FileData
 from vorta.views.partials.treemodel import FileSystemItem, path_to_str
 
 from .borg_job import BorgJob
