+++ src/vorta/views/dialogs/exception_dialog.py
@@ -7,10 +7,8 @@ from PyQt6.QtWidgets import QApplication
 from vorta._version import __version__
 from vorta.i18n import trans_late, translate
 from vorta.i18n.richtext import escape, format_richtext, link
-from vorta.utils import borg_compat
+from vorta.utils import borg_compat, get_asset
 from vorta.views.utils import get_colored_icon
-
-from ..utils import get_asset
 
 # Load UI file
 uifile = get_asset('UI/exception_dialog.ui')
