+++ src/vorta/views/dialogs/export_window.py
@@ -7,11 +7,10 @@ from PyQt6.QtCore import Qt
 from PyQt6.QtWidgets import QFileDialog, QMessageBox
 
 from vorta.keyring.abc import VortaKeyring
+from vorta.notifications import VortaNotifications
+from vorta.profile_export import ProfileExport
 from vorta.store.models import BackupProfileModel  # noqa: F401
 from vorta.utils import get_asset
-
-from ..notifications import VortaNotifications
-from ..profile_export import ProfileExport
 
 uifile_import = get_asset('UI/export_window.ui')
 ExportWindowUI, ExportWindowBase = uic.loadUiType(uifile_import)
