+++ src/vorta/views/repo_tab.py
@@ -11,9 +11,9 @@ from vorta.store.models import ArchiveModel, RepoModel
 from vorta.utils import borg_compat, get_asset, get_private_keys, pretty_bytes
 
 from .base_tab import BaseTab
-from .repo_add_dialog import AddRepoWindow, ExistingRepoWindow
-from .repo_change_passphrase import ChangeBorgPassphraseWindow
-from .ssh_dialog import SSHAddWindow
+from .dialogs.repo_add_dialog import AddRepoWindow, ExistingRepoWindow
+from .dialogs.repo_change_passphrase import ChangeBorgPassphraseWindow
+from .dialogs.ssh_dialog import SSHAddWindow
 from .utils import get_colored_icon
 
 uifile = get_asset('UI/repo_tab.ui')
