+++ tests/unit/test_profile.py
@@ -4,7 +4,7 @@ from PyQt6 import QtCore
 from PyQt6.QtWidgets import QDialogButtonBox, QFileDialog, QMessageBox, QToolTip
 
 from vorta.store.models import BackupProfileModel, SourceFileModel
-from vorta.views.export_window import ExportWindow
+from vorta.views.dialogs.export_window import ExportWindow
 
 
 def test_profile_add_delete(qapp, qtbot, mocker):
